Backend/requirements.txt
fastapi 
uvicorn 
pymongo 
python-dotenv 
pydantic
flask
flask_cors